# api/openapi.yaml
openapi: 3.0.3
info:
  title: 家庭健康管理 API
  version: 0.2.0
  description: |
    接口契约的唯一事实来源。Go 服务端接口（oapi-codegen）和前端 TS 类型（openapi-typescript）
    都由本文件生成，修改后执行 `make gen`。

    约定（详见 docs/tech-plan.md「API 设计」）：
    - 业务接口统一前缀 `/api`，JSON 字段用 camelCase。
    - 时间用 ISO 8601 带时区偏移（2026-09-24T21:30:00+08:00），纯日期用 2026-09-24；按天统计以 Asia/Shanghai 为准。
    - 记录和附件用 `PUT /{id}` 幂等创建，ID 由前端生成（UUIDv7）：首次创建返回 201，重复提交返回 200 和已有资源。
    - PATCH 只改请求里出现的字段；可清空的字段传 null 表示清空。
    - 列表用游标分页，默认每页 50 条。
    - 鉴权用 Cookie `sid`（登录接口下发）；所有写请求校验 Origin 头。
    - 错误统一返回 Error 结构。
servers:
  - url: /
security:
  - cookieAuth: []
...
paths:
...
  /api/attachments/{id}:
    parameters:
      - $ref: "#/components/parameters/Id"
    put:
      tags: [attachments]
      operationId: uploadAttachment
      summary: 上传附件（幂等）。照片和头像须为 JPEG；语音上传后异步转码
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema: { $ref: "#/components/schemas/AttachmentUpload" }
      responses:
        "200":
          description: 已存在
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Attachment" }
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Attachment" }
        default: { $ref: "#/components/responses/Error" }
    patch:
      tags: [attachments]
      operationId: updateAttachment
      summary: 修改语音补充文字或顺序
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/AttachmentPatch" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Attachment" }
        default: { $ref: "#/components/responses/Error" }
    delete:
      tags: [attachments]
      operationId: deleteAttachment
      responses:
        "204": { description: 已删除 }
        default: { $ref: "#/components/responses/Error" }
  /api/attachments/{id}/file:
    parameters:
      - $ref: "#/components/parameters/Id"
    get:
      tags: [attachments]
      operationId: getAttachmentFile
      summary: 读取附件文件，支持 Range（语音可拖动）
      description: |
        登录态或打印令牌（token）二选一。语音默认返回转码后的 m4a，未转码完成时返回原始文件。
      security: []
      parameters:
        - name: variant
          in: query
          schema:
            type: string
            enum: [default, thumb, original]
            default: default
        - name: token
          in: query
          description: 打印令牌，供 Gotenberg 渲染报告时取图
          schema: { type: string }
      responses:
        "200":
          description: 文件内容
          content:
            application/octet-stream:
              schema: { type: string, format: binary }
        "206":
          description: 部分内容（Range）
          content:
            application/octet-stream:
              schema: { type: string, format: binary }
        default: { $ref: "#/components/responses/Error" }
  /api/attachments/{id}/reprocess:
    parameters:
      - $ref: "#/components/parameters/Id"
    post:
      tags: [attachments]
      operationId: reprocessAttachment
      summary: 转码失败的语音重新处理
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Attachment" }
        default: { $ref: "#/components/responses/Error" }
...
  /api/print-data:
    get:
      tags: [reports]
      operationId: getPrintData
      summary: 打印页取数
      description: |
        Gotenberg 渲染时带一次性打印令牌（token）；导出页预览和浏览器打印凭登录态，并用 type/id/from/to/photos 指定范围。
      security: []
      parameters:
        - name: token
          in: query
          schema: { type: string }
        - name: type
          in: query
          schema: { $ref: "#/components/schemas/ReportType" }
        - name: id
          in: query
          description: 病程报告为病程 ID，成员健康档案为成员 ID
          schema: { type: string, format: uuid }
        - name: from
          in: query
          schema: { type: string, format: date }
        - name: to
          in: query
          schema: { type: string, format: date }
        - name: photos
          in: query
          schema: { $ref: "#/components/schemas/PhotoOption" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/PrintData" }
        default: { $ref: "#/components/responses/Error" }
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: sid
  parameters:
    Id:
      name: id
      in: path
      required: true
      schema: { type: string, format: uuid }
...
  responses:
    Error:
      description: 错误（400/401/403/404/409/413/422/429/500）
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
  schemas:
...
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
              description: 机器可读的错误码，如 validation_failed、not_found、unauthorized
              example: validation_failed
            message:
              type: string
              description: 可直接展示给用户的中文说明
              example: 体温需在 34.0 到 43.0 之间
            fields:
              type: object
              description: 字段级错误，键为字段名
              additionalProperties: { type: string }
              example: { temperature: out_of_range }

    # ---------- 枚举 ----------
    MemberRelation:
      type: string
      enum: [self, spouse, child, parent, grandparent, other]
    Gender:
      type: string
      enum: [male, female]
    BloodType:
      type: string
      enum: [A, B, AB, O, unknown]
    EpisodeKind:
      type: string
      description: short 短期病程，long 长期病程
      enum: [short, long]
    EpisodeStatus:
      type: string
      description: 短期：active 进行中、recovered 已康复；长期：treating 治疗中、stable 稳定期、ended 已结束
      enum: [active, recovered, treating, stable, ended]
    RecordType:
      type: string
      enum: [symptom, temperature, medication, visit, treatment, exam, other]
    MedUnit:
      type: string
      enum: [ml, 片, 粒, 袋]
      x-enum-varnames: [Ml, Tablet, Capsule, Sachet]
    AttachmentKind:
      type: string
      enum: [photo, audio, avatar]
    AttachmentStatus:
      type: string
      description: 照片和头像上传后即为 ready；语音为 processing，转码完成后 ready，三次失败为 failed
      enum: [pending, processing, ready, failed]
    ReportType:
      type: string
      description: episode 病程报告，member 成员健康档案
      enum: [episode, member]
    PhotoOption:
      type: string
      description: none 不附带，thumbnail 附缩略图，appendix 作为附录附原图
      enum: [none, thumbnail, appendix]

    # ---------- 账号 ----------
...
    Member:
      type: object
      required: [id, nickname, relation, gender, birthDate, avatarId, avatarUrl, allergies, bloodType, notes, sortOrder, archived, longEpisodes, createdAt, updatedAt]
      properties:
        id: { type: string, format: uuid }
        nickname: { type: string }
        relation: { $ref: "#/components/schemas/MemberRelation" }
        gender: { $ref: "#/components/schemas/Gender" }
        birthDate: { type: string, format: date }
        avatarId: { type: string, format: uuid, nullable: true }
        avatarUrl: { type: string, nullable: true }
        allergies: { type: string, nullable: true }
        bloodType:
          allOf: [{ $ref: "#/components/schemas/BloodType" }]
          nullable: true
        notes: { type: string, nullable: true }
        sortOrder: { type: integer }
        archived: { type: boolean }
        longEpisodes:
          description: 该成员的长期病程（长期疾病由此汇总显示）
          type: array
          items: { $ref: "#/components/schemas/EpisodeRef" }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
...
    EpisodeRef:
      type: object
      required: [id, name, diseaseName, kind, status]
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        diseaseName: { type: string }
        kind: { $ref: "#/components/schemas/EpisodeKind" }
        status: { $ref: "#/components/schemas/EpisodeStatus" }
    Episode:
      type: object
      required: [id, memberId, diseaseTagId, diseaseName, name, kind, status, open, startedOn, endedOn, days, recordCount, lastRecordAt, costTotalCents, suggestRecovered, createdAt, updatedAt]
      properties:
        id: { type: string, format: uuid }
        memberId: { type: string, format: uuid }
        diseaseTagId: { type: string, format: uuid }
        diseaseName: { type: string }
        name: { type: string }
        kind: { $ref: "#/components/schemas/EpisodeKind" }
        status: { $ref: "#/components/schemas/EpisodeStatus" }
        open: { type: boolean, description: 未结束（进行中、治疗中、稳定期） }
        startedOn: { type: string, format: date }
        endedOn: { type: string, format: date, nullable: true }
        days: { type: integer, description: 第几天（未结束）或共持续几天（已结束） }
        recordCount: { type: integer }
        lastRecordAt: { type: string, format: date-time, nullable: true }
        costTotalCents: { type: integer, description: 就诊与治疗费用合计，单位分 }
        suggestRecovered:
          type: boolean
          description: 短期病程超过 14 天没有新记录，打开时提示“是否已康复”
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
...
    RecordDetails:
      type: object
      description: 就诊：hospital、department、doctor；治疗：item、institution；检查：item
      properties:
        hospital: { type: string, maxLength: 100 }
        department: { type: string, maxLength: 50 }
        doctor: { type: string, maxLength: 50 }
        item: { type: string, maxLength: 100 }
        institution: { type: string, maxLength: 100 }
    Record:
      type: object
      required: [id, memberId, episodeId, type, occurredAt, createdAt, updatedAt, backfilled, body, isFlare, severity, temperature, medName, medDose, medUnit, costCents, details, attachments]
      properties:
        id: { type: string, format: uuid }
        memberId: { type: string, format: uuid }
        episodeId: { type: string, format: uuid, nullable: true }
        type:
          allOf: [{ $ref: "#/components/schemas/RecordType" }]
          nullable: true
        occurredAt: { type: string, format: date-time, description: 发生时间 }
        createdAt: { type: string, format: date-time, description: 录入时间 }
        updatedAt: { type: string, format: date-time }
        backfilled: { type: boolean, description: 发生与录入相差超过 1 小时，显示“补录” }
        body: { type: string }
        isFlare: { type: boolean }
        severity: { type: integer, nullable: true }
        temperature: { type: number, format: double, nullable: true }
        medName: { type: string, nullable: true }
        medDose: { type: number, format: double, nullable: true }
        medUnit:
          allOf: [{ $ref: "#/components/schemas/MedUnit" }]
          nullable: true
        costCents: { type: integer, nullable: true }
        details: { $ref: "#/components/schemas/RecordDetails" }
        attachments:
          type: array
          items: { $ref: "#/components/schemas/Attachment" }
...
    Attachment:
      type: object
      required: [id, recordId, kind, status, mime, sizeBytes, durationMs, width, height, caption, sortOrder, url, thumbUrl, createdAt]
      properties:
        id: { type: string, format: uuid }
        recordId: { type: string, format: uuid, nullable: true }
        kind: { $ref: "#/components/schemas/AttachmentKind" }
        status: { $ref: "#/components/schemas/AttachmentStatus" }
        mime: { type: string }
        sizeBytes: { type: integer, format: int64 }
        durationMs: { type: integer, nullable: true }
        width: { type: integer, nullable: true }
        height: { type: integer, nullable: true }
        caption: { type: string, nullable: true, description: 语音的补充文字 }
        sortOrder: { type: integer }
        url: { type: string }
        thumbUrl: { type: string, nullable: true }
        createdAt: { type: string, format: date-time }
    AttachmentUpload:
      type: object
      required: [kind, file]
      properties:
        kind: { $ref: "#/components/schemas/AttachmentKind" }
        recordId: { type: string, format: uuid, description: 照片和语音必填，头像不填 }
        durationMs: { type: integer, minimum: 0, maximum: 180000, description: 前端测得的语音时长，转码后以 ffprobe 为准 }
        caption: { type: string, maxLength: 200 }
        sortOrder: { type: integer }
        file:
          type: string
          format: binary
          description: 照片 image/jpeg ≤10MB；语音 audio/mp4、audio/webm、audio/ogg ≤20MB；头像 image/jpeg ≤2MB
    AttachmentPatch:
      type: object
      properties:
        caption:
          type: string
          maxLength: 200
          nullable: true
          x-go-type: nullable.Nullable[string]
          x-go-type-import: { path: github.com/oapi-codegen/nullable }
          x-go-type-skip-optional-pointer: true
        sortOrder: { type: integer }

    # ---------- 浏览 ----------
...
    PrintData:
      type: object
      required: [type, generatedAt, from, to, photos, member, episode, episodes, records, costTotalCents, token]
      properties:
        type: { $ref: "#/components/schemas/ReportType" }
        generatedAt: { type: string, format: date-time }
        from: { type: string, format: date, nullable: true }
        to: { type: string, format: date, nullable: true }
        photos: { $ref: "#/components/schemas/PhotoOption" }
        member: { $ref: "#/components/schemas/Member" }
        episode:
          allOf: [{ $ref: "#/components/schemas/Episode" }]
          nullable: true
          description: 病程报告的病程
        episodes:
          description: 成员健康档案中时间段内的病程（每个一行摘要）
          type: array
          items: { $ref: "#/components/schemas/Episode" }
        records:
          description: 病程报告中时间段内的全部记录，按发生时间升序
          type: array
          items: { $ref: "#/components/schemas/Record" }
        costTotalCents: { type: integer }
        token:
          type: string
          nullable: true
          description: 取数用的打印令牌，打印页取附件文件时附在 URL 上